bugfixes for krun options
Former-commit-id: 2417256dea088540c6694aebdfcb6a31ffe72369 [formerly 40b3951bc720d3882bbeea05dac8686487fa2082]
Former-commit-id: 8ab7ada4e02ec9004e5454cb6788b5c7b0b2354f

+++ src/javasources/KTool/src/org/kframework/backend/latex/PdfBackend.java
@@ -63,6 +63,7 @@ public class PdfBackend extends BasicBackend {
         latexBackend.compile(definition);
         File latexFile = latexBackend.getLatexFile();
         File pdfFile = generatePdf(latexFile);
+        if (pdfFile.exists()) {
             copyFile(pdfFile, options.output());
         }
     }
